minor changes (mainly header rename) for lib overhaul
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@157 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/CStr.h
@@ -173,7 +173,7 @@ public:
 	TCHAR &operator[](_ulong n);
 
 	// Serialization functions
-	virtual uint GetSerializedLength() const;
+	virtual size_t GetSerializedLength() const;
 	virtual u8 *Serialize(u8 *buffer) const;
 	virtual const u8 *Deserialize(const u8 *buffer, const u8 *bufferend);
